refactor: request, response 역할에 따른 분리
