+++ src/main/java/webserver/request/RequestHeaderParser.java
@@ -1,9 +1,9 @@
-package webserver;
+package webserver.request;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class HttpHeaderParser {
+public class RequestHeaderParser {
     static String isLogin(String cookieString) {
         if (cookieString.isEmpty()) {
             return "false";
