+++ src/main/java/webserver/request/UserCreate.java
@@ -1,7 +1,8 @@
-package webserver;
+package webserver.request;
 
 import db.DataBase;
 import model.User;
+import webserver.request.QueryStringParser;
 
 import java.util.Map;
