+++ src/main/resources/templates/user/list.html
@@ -93,7 +93,7 @@
                 </thead>
                 {{#each this}}
                 <tr>
-                    <th scope="row">1</th>
+                    <th scope="row">{{@index}}</th>
                     <td>{{userId}}</td>
                     <td>{{name}}</td>
                     <td>{{email}}</td>
