+++ src/test/java/webserver/HttpHeaderParserTest.java
@@ -2,6 +2,7 @@ package webserver;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.request.RequestHeaderParser;
 
 import static org.assertj.core.api.Assertions.assertThat;
