+++ src/test/java/webserver/RequestLineTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpMethod;
+import webserver.request.RequestLine;
 
 public class RequestLineTest {
     @Test
